Allow for queries with order by but no limit (#1545)
This unifies our serialization of K to always use an int (as we're
limited by Go slice lengths anyway) where -1 means no limit.

+++ go/boost/server/controller/controller.go
@@ -263,7 +263,7 @@ func (ctrl *Controller) viewDescriptor(node *flownode.Node) *vtboostpb.Materiali
 		Shards:        shards,
 		TopkOrderCols: orderCols,
 		TopkOrderDesc: orderColsDesc,
-		TopkLimit:     orderLimit,
+		TopkLimit:     int64(orderLimit),
 	}
 }
